Correcciones a tests
