+++ profile/apps.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+
+from django.apps import AppConfig
+
+
+class ProfileConfig(AppConfig):
+    name = 'profile'
